requirements.txt
flask
python-dotenv
requests
jinja2